+++ src/main/java/com/ratemyschools/rate/dto/ClubReviews/GetClubsAverageRatingDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class GetClubsAverageRatingDto {
     private Long id;
-    private Integer averageRating;
+    private Double averageRating;
 
     public Long getId() {
         return id;
@@ -17,11 +17,11 @@ public class GetClubsAverageRatingDto {
         this.id = id;
     }
 
-    public Integer getAverageRating() {
+    public Double getAverageRating() {
         return averageRating;
     }
 
-    public void setAverageRating(Integer averageRating) {
+    public void setAverageRating(Double averageRating) {
         this.averageRating = averageRating;
     }
 }
